requirements.txt
numpy==1.22.0
opencv-python==4.2.0.32
tensorflow==2.9.3
flask>=2.0
flask-cors>=3.0
Pillow>=9.0
